backend/requirements.txt
fastapi==0.135.1
uvicorn[standard]==0.41.0
python-multipart==0.0.22
python-dotenv==1.2.2
supabase==2.28.0
SQLAlchemy[asyncio]==2.0.48
asyncpg==0.31.0
shapely==2.1.2
earthengine-api==1.7.17
asf_search==7.0.5
numpy==2.4.3
scipy==1.13.0
paddlepaddle==3.3.0
paddleocr==3.4.0
opencv-python-headless==4.10.0.84
Pillow==12.1.1
xgboost==2.0.3
scikit-learn==1.4.2
celery==5.6.2
flower==2.0.1
redis==7.3.0
web3==7.14.1
eth-account==0.13.7
requests==2.32.5
httpx==0.28.1
pydantic==2.12.5
pydantic-settings==2.13.1
lru-dict==1.2.0